Own module for player.
