+++ src/main.js
@@ -21,7 +21,8 @@
  * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-import { init, Sprite, GameLoop, bindKeys, initKeys, keyPressed } from "kontra";
+import { init, Sprite, GameLoop, bindKeys, initKeys } from "kontra";
+import { createPlayer } from "./player.js";
 import playerSvg from "./images/player.svg";
 import houseSvg from "./images/house.svg";
 
@@ -32,15 +33,6 @@ const imageFromSvg = svgString => {
   image.src = base64Header + svgInBase64;
   return image;
 };
-
-const playerSpeed = 3;
-const gravity = 2;
-const jumpVelocity = -30;
-const climbSpeed = 2;
-
-const STATE_ON_GROUND = 0;
-const STATE_JUMPING = 1;
-const STATE_CLIMBING = 2;
 
 const CAMERA_MODE_FOLLOW_PLAYER = 0;
 const CAMERA_MODE_SHOW_WHOLE_LEVEL = 1;
@@ -129,6 +121,17 @@ let camera = {
   }
 };
 
+const isPlayerOnLadders = () => {
+  for (let i = 0; i < ladders.length; i++) {
+    let ladder = ladders[i];
+    if (ladder.collidesWith(player)) {
+      return true;
+    }
+  }
+
+  return false;
+};
+
 let loop = GameLoop({
   update: () => {
     for (let i = 0; i < clouds0.length; i++) {
@@ -136,7 +139,7 @@ let loop = GameLoop({
       clouds1[i].update();
     }
 
-    player.update();
+    player.update(isPlayerOnLadders());
 
     camera.update();
   },
@@ -355,78 +358,6 @@ const createCloud = (y, z) => {
   }
 };
 
-const createPlayer = () => {
-  return Sprite({
-    color: "red",
-    width: 50,
-    height: 150,
-    vel: 0, // Vertical velocity, affected by jumping and gravity
-    state: STATE_ON_GROUND,
-
-    isOnGround: function() {
-      const margin = 5;
-      return this.y + this.height > level.height - margin;
-    },
-
-    render: function() {
-      this.context.drawImage(playerImage, this.x, this.y);
-    },
-
-    update: function() {
-      let dx = 0;
-      let dy = 0;
-
-      if (keyPressed("left") && this.x > 0) {
-        dx = -playerSpeed;
-      } else if (keyPressed("right") && this.x < level.width - this.width) {
-        dx = playerSpeed;
-      }
-
-      let canClimb = false;
-      for (let i = 0; i < ladders.length; i++) {
-        let ladder = ladders[i];
-        if (ladder.collidesWith(player)) {
-          canClimb = true;
-        }
-      }
-
-      if (!canClimb && this.state === STATE_CLIMBING) {
-        this.state = STATE_ON_GROUND;
-      }
-
-      if (keyPressed("up")) {
-        if (canClimb) {
-          this.state = STATE_CLIMBING;
-          this.vel = 0;
-          dy -= climbSpeed;
-        } else if (this.state !== STATE_JUMPING && this.isOnGround()) {
-          this.vel = jumpVelocity;
-          this.state = STATE_JUMPING;
-        }
-      } else if (keyPressed("down") && canClimb) {
-        this.state = STATE_CLIMBING;
-        this.vel = 0;
-        dy += climbSpeed;
-      }
-
-      if (this.state !== STATE_CLIMBING) {
-        this.vel += gravity;
-        dy += this.vel;
-      }
-
-      this.x += dx;
-
-      if (this.y + dy > level.height - this.height) {
-        this.y = level.height - this.height;
-        this.vel = 0;
-        this.state = STATE_ON_GROUND;
-      } else {
-        this.y += dy;
-      }
-    }
-  });
-};
-
 const createLadder = () => {
   return Sprite({
     color: "gray",
@@ -487,7 +418,7 @@ const initScene = () => {
   ladder.y = level.top;
   ladders.push(ladder);
 
-  player = createPlayer();
+  player = createPlayer(level, playerImage);
   player.x = 30;
   player.y = level.height / 2 - player.height;
 };
